+++ operator/pkg/reconciliation/construct_service_test.go
@@ -1,0 +1,29 @@
+// Copyright DataStax, Inc.
+// Please see the included license file for details.
+
+package reconciliation
+
+import (
+	"reflect"
+	"testing"
+
+	api "github.com/datastax/cass-operator/operator/pkg/apis/cassandra/v1beta1"
+)
+
+func TestCassandraDatacenter_buildLabelSelectorForSeedService(t *testing.T) {
+	dc := &api.CassandraDatacenter{
+		Spec: api.CassandraDatacenterSpec{
+			ClusterName: "bob",
+		},
+	}
+	want := map[string]string{
+		api.ClusterLabel:  "bob",
+		api.SeedNodeLabel: "true",
+	}
+
+	got := buildLabelSelectorForSeedService(dc)
+
+	if !reflect.DeepEqual(want, got) {
+		t.Errorf("buildLabelSelectorForSeedService = %v, want %v", got, want)
+	}
+}
